product list some error fixed
