--- client/src/Pages/Products/AddProduct.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const AddProduct = () => {
-  return (
-    <div className=" max-w-screen-xl mx-auto lg:px-4 font-poppins">
-      AddProduct
-    </div>
-  );
-};
-
-export default AddProduct;
