+++ server/index.js
@@ -28,6 +28,8 @@ app.use(cors());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
+app.use(express.static("images"));
+
 // Middleware to parse cookies
 app.use(cookieParser());
